Add new interfaces 'Message', 'FileRequest' and 'FileResponse' to make my code more generalized and flexible.
