+++ common/src/main/java/ru/ryabtsev/cloud/common/message/server/HandshakeResponse.java
