+++ common/src/main/java/ru/ryabtsev/cloud/common/message/server/file/FileResponse.java
@@ -1,0 +1,4 @@
+package ru.ryabtsev.cloud.common.message.server.file;
+
+public interface FileResponse {
+}
